% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table.R
\name{aggregate_candidate_votes}
\alias{aggregate_candidate_votes}
\title{Create data frame of results of election by state}
\usage{
aggregate_candidate_votes(votes, keep_percentage = 0.01)
}
\arguments{
\item{votes}{The votes data frame for the elections}

\item{keep_percentage}{What percentage of the votes should a candidate have
gotten in order to be displayed in the data frame? In othe words, how much
of a contender did the person have to be?}
}
\description{
This creates a data frame of results for each Congressional district in a
state.
}
\examples{
meae_id <- "meae.congressional.congress08.md.county"
votes <- vote_counts(meae_id)
aggregate_candidate_votes(votes)
}
