% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{aggregate_party_votes}
\alias{aggregate_party_votes}
\title{Aggregate votes for a geographic unity by political party}
\usage{
aggregate_party_votes(data, geography = c("county"),
  parties = c("Federalist", "Republican"))
}
\arguments{
\item{data}{A data frame returned by \code{\link{vote_counts}}.}

\item{geography}{The geography to aggregate by for mapping. For instance,
\code{"county"}.}

\item{parties}{The parties to explicitly keep by name. All other parties,
along with unaffiliated candidates, will be collapsed into \code{"other"}.}
}
\value{
A data frame with one row for each geographic unit (e.g. for each
  county) in the data frame. There will be columns for each party, giving
  their votes recorded and vote percentage, along with the total vote in that
  geographic unit.
}
\description{
Aggregate votes for a geographic unity by political party
}
\examples{
votes <- vote_counts("meae.congressional.congress05.ny.county")
aggregate_party_votes(votes)

}
