% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-yaml.R
\name{generate_map_yaml}
\alias{generate_map_yaml}
\title{Generate the YAML for each Congressional election map}
\usage{
generate_map_yaml(meae_id, legend_type = "fed-vs-repub-percentage",
  always_allow_html = "true", layout = "maps-show")
}
\arguments{
\item{meae_id}{A unique Mapping Early American Elections id.}

\item{legend_type}{A description of the data visualization. Default set to fed-vs-reub-percentage.}

\item{always_allow_html}{Default to TRUE.}

\item{layout}{Type of html layout. Default set to maps-show.}
}
\value{
A list of all the needed metadata for rendering the maps in Jekyll
}
\description{
Generate the YAML for each Congressional election map
}
