% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{get_national_map_data}
\alias{get_national_map_data}
\title{Get the data to map party vote percentages by county for the nation}
\usage{
get_national_map_data(map_id)
}
\arguments{
\item{map_id}{The ID of the map from \code{meae_maps}.}
}
\description{
Get the data to map party vote percentages by county for the nation
}
\examples{
get_national_map_data("meae.congressional.congress01.national.county")
}
