% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{get_staterep_map_data}
\alias{get_staterep_map_data}
\title{Get the spatial data to map party vote percentages by county for state legislative elections}
\usage{
get_staterep_map_data(map_id, state = NULL, election_date = NULL)
}
\arguments{
\item{map_id}{The ID of the map from \code{meae_maps}.}

\item{state}{The state that the map should display.}

\item{election_date}{The date to be used when getting the county boundaries.}
}
\value{
An sf spatial data frame joined to the election data.
}
\description{
Get the spatial data to map party vote percentages by county for state legislative elections
}
\examples{
get_staterep_map_data("meae.staterepresentative.1796.de.county", state = "DE",
                      election_date = "1796-10-01")
}
