% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{histcongress}
\alias{histcongress}
\title{Historic Congressional district boundaries}
\format{sf polygons object}
\source{
Jeffrey B. Lewis, Brandon DeVine, Lincoln Pitcher, and Kenneth C. Martis. (2013) Digital Boundary Definitions of United States Congressional Districts, 1789-2012. [Data file and code book]. Retrieved from \url{http://cdmaps.polisci.ucla.edu}.
}
\usage{
histcongress
}
\description{
Historic Congressional district boundaries
}
\keyword{datasets}
