% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{join_to_spatial}
\alias{join_to_spatial}
\title{Join aggregated party data to spatial data}
\usage{
join_to_spatial(data, geography = c("county"), resolution = c("high",
  "low"), date = NULL)
}
\arguments{
\item{data}{A data frame returned by \code{\link{aggregate_party_votes}}.}

\item{geography}{The geography to aggregate by for mapping. For instance,
\code{"county"}.}

\item{resolution}{Use \code{"high"} or \code{"low"} resolution geographic
data.}

\item{date}{Override the date for the geographies used in this map. Otherwise
the appropriate date will be guessed from the data passed in.}
}
\value{
A \code{sf} object with polygons for the appropriate geographies and
  the data joined to them.
}
\description{
Join aggregated party data to spatial data
}
\examples{
votes <- vote_counts("meae.congressional.congress05.ny.county")
aggregates <- aggregate_party_votes(votes)
join_to_spatial(aggregates, resolution = "low")

}
