% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template.R
\name{make_map_page}
\alias{make_map_page}
\title{Make a map page for the Jekyll site}
\usage{
make_map_page(meae_id, ..., dir, overwrite = FALSE)
}
\arguments{
\item{meae_id}{The MEAE ID of the map to be created.}

\item{...}{Additional arguments passed on to
\code{\link{generate_map_metadata}}.}

\item{dir}{The directory in which to create R Markdown file. The file will be
named after the \code{meae_id}.}

\item{overwrite}{If the file already exists, should it be overwritten?}
}
\value{
Invisibly returns the path to the file that was created.
}
\description{
This function generates an R Markdown page which in turn generates the map
view page in the Jekyll website. The R Markdown file will be created in the
directory passed to the function (presumably the \code{_maps}) directory in
Jekyll). It contains the YAML metadata header plus the R code that creates
the maps and tables. This file will need to be edited by hand to include text
and captions, so by default this function will not overwrite the created
file. The R Markdown file has to be run separately to create the Markdown
file used by Jekyll.
}
\examples{
meae_id <- "meae.congressional.congress08.va.county"
# The function invisibly returns the path to the file it created
file <- make_map_page(meae_id, dir = tempdir(), overwrite = TRUE)
cat(readr::read_file(file))
}
