% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\docType{data}
\name{map_elections}
\alias{map_elections}
\alias{federalist_vs_republican}
\title{Map elections data}
\format{An object of class \code{list} of length 3.}
\usage{
map_elections(data, projection = NULL, legend = FALSE,
  congressional_boundaries = TRUE, state_boundaries = FALSE, cities = 8L,
  scale = federalist_vs_republican, width = "100\%", height = NULL)

federalist_vs_republican
}
\arguments{
\item{data}{An \code{sf} object with elections data returned by
\code{\link{join_to_spatial}}.}

\item{projection}{If not provided, then the best state plane projection will
be guessed using the \code{\link[USAboundaries]{state_plane}} function from
the \code{USAboundaries} package. If \code{NULL}, then leaflet's default
Web Mercator projection will be used. To use a custom projection, provide a
projection/CRS object returned by the \code{\link[leaflet]{leafletCRS}}
function in the \code{leaflet} package.}

\item{legend}{Should a legend be displayed or not?}

\item{congressional_boundaries}{Draw Congressional district boundaries in
addition to county boundaries?}

\item{state_boundaries}{Draw state boundaries in addition to county
boundaries?}

\item{cities}{Number of largest cities to draw. Pass \code{FALSE} to not draw
any cities.}

\item{scale}{The type of scale to use for the choropleth map.}

\item{width}{The width of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}

\item{height}{The height of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}
}
\description{
Map elections data
}
\examples{
meae_id <- "meae.congressional.congress08.va.county"
votes <- vote_counts(meae_id)
aggregates <- aggregate_party_votes(votes)
map_data <- join_to_spatial(aggregates)
map_elections(map_data, legend = TRUE)

}
\keyword{datasets}
