% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_national}
\alias{map_national}
\title{Map the Congressional data for the nation}
\usage{
map_national(data, congress = NULL, congressional_boundaries = TRUE,
  state_boundaries = FALSE, width = "100\%", height = "800px")
}
\arguments{
\item{data}{An \code{sf} object with elections data returned by
\code{\link{join_to_spatial}}.}

\item{congress}{The number of the Congress. If \code{NULL}, it will be
guessed from the data.}

\item{congressional_boundaries}{Draw Congressional district boundaries in
addition to county boundaries?}

\item{state_boundaries}{Draw state boundaries in addition to county
boundaries?}

\item{width}{The width of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}

\item{height}{The height of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}
}
\description{
Map the Congressional data for the nation
}
\examples{
map_data <- get_national_map_data("meae.congressional.congress01.national.county")
map_national(map_data)
}
