% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{meae_candidates}
\alias{meae_candidates}
\alias{meae_congressional_counties}
\alias{meae_elections}
\title{Elections data from Mapping Early American Elections}
\format{Data frames}
\usage{
meae_candidates

meae_congressional_counties

meae_elections
}
\description{
Elections data from Mapping Early American Elections
}
\keyword{datasets}
