% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-nnv.R
\name{parse_nnv_xml}
\alias{parse_nnv_xml}
\title{Parse an election XML file from NNV}
\usage{
parse_nnv_xml(file)
}
\arguments{
\item{file}{Path to the XML file.}
}
\description{
Given the path to an NNV file with election results in NNV, return the totals
by geography for each candidate.
}
\examples{
\dontrun{
file1 <- "~/dev/mapping-elections/nnv-xml/New York/1789/ny.uscongress1.1789.xml"
file2 <- "~/dev/mapping-elections/nnv-xml/Virginia/1789/va.uscongress.7.1789.xml"
file3 <- "~/dev/mapping-elections/nnv-xml/Tennessee/1801/tn.congress.1801.xml"
parse_nnv_xml(file1)
parse_nnv_xml(file2)
parse_nnv_xml(file3)
}
}
