% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table.R
\name{candidate_results_state_rep}
\alias{candidate_results_state_rep}
\alias{results_to_table_state_rep}
\title{Get a table of results by candidate and district for state legislative representatives}
\usage{
candidate_results_state_rep(map_id)

results_to_table_state_rep(results, keep_percentage = 0.05)
}
\arguments{
\item{map_id}{An ID from the \code{meae_maps} table.}

\item{results}{A data frame of results from \code{candidate_results}.}

\item{keep_percentage}{What percentage of votes defines a contender?}
}
\description{
Get a table of results by candidate and district for state legislative representatives

Format candidate results in an HTML table for state legislative representatives
}
\examples{
map_id <- "meae.staterepresentative.1796.de.county"
candidate_results_state_rep(map_id)
map_id <- "meae.staterepresentative.1796.de.county"
results <- candidate_results_state_rep(map_id)
results_to_table_state_rep(results)
}
