% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table.R
\name{candidate_results}
\alias{candidate_results}
\alias{results_to_table}
\title{Get a table of results by candidate and Congressional district}
\usage{
candidate_results(map_id)

results_to_table(results)
}
\arguments{
\item{map_id}{An ID from the \code{meae_maps} table.}

\item{results}{A data frame of results from \code{candidate_results}.}
}
\description{
Get a table of results by candidate and Congressional district

Format candidate results in an HTML table
}
\examples{
map_id <- "meae.congressional.congress01.ny.county"
candidate_results(map_id)
map_id <- "meae.congressional.congress01.pa.county"
results <- candidate_results(map_id)
results_to_table(results)
}
