% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results-table.R
\name{results_to_table}
\alias{results_to_table}
\title{Create table of results for election}
\usage{
results_to_table(results)
}
\description{
Create table of results for election
}
\examples{
meae_id <- "meae.congressional.congress08.md.county"
votes <- vote_counts(meae_id)
results <- aggregate_candidate_votes(votes)
results_to_table(results)
}
