% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{vote_counts}
\alias{vote_counts}
\title{Get the votes by candidate and geographic unit}
\usage{
vote_counts(map_id)
}
\arguments{
\item{map_id}{An MEAE map ID from the \code{\link{meae_maps}} table.}
}
\value{
A data frame containing the votes for each candidate in a county.
}
\description{
Given a MEAE map ID, get the votes by candidate and geographic unit for the
elections associated with that map.
}
\examples{
vote_counts("meae.congressional.congress08.md.county")

}
