% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-yaml.R
\name{generate_map_metadata}
\alias{generate_map_metadata}
\title{Generate the YAML for each Congressional election map}
\usage{
generate_map_metadata(meae_id, legend = "fed-vs-anti",
  always_allow_html = TRUE, layout_type = "maps-show")
}
\arguments{
\item{meae_id}{A unique Mapping Early American Elections id.}

\item{always_allow_html}{Default to TRUE.}

\item{layout_type}{Type of html layout. Default set to maps-show.}
}
\value{
A list of all the needed metadata for rendering the maps in Jekyll
}
\description{
Generate the YAML for each Congressional election map
}
\examples{
generate_map_metadata(meae_id = "meae.congressional.congress05.ny.county")
}
