% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-manipulation.R
\name{join_to_spatial}
\alias{join_to_spatial}
\title{Join aggregated party data to spatial data}
\usage{
join_to_spatial(party_votes, elections, resolution = c("high", "low"),
  election_date = NULL)
}
\arguments{
\item{party_votes}{A data frame of vote percentages by party.}

\item{elections}{A data frame containing metadata about the NNV elections
being mapped.}

\item{resolution}{Use \code{"high"} or \code{"low"} resolution geographic
data.}

\item{election_date}{The date to use when getting the county boundaries.}
}
\value{
A \code{sf} object with polygons for the appropriate geographies and
  the data joined to them.
}
\description{
Join aggregated party data to spatial data
}
