% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{map_counties}
\alias{map_counties}
\title{Map elections data}
\usage{
map_counties(data, congress = NULL, projection = NULL,
  congressional_boundaries = TRUE, cities = 4L, width = "100\%",
  height = "600px")
}
\arguments{
\item{data}{An \code{sf} object with elections data returned by
\code{\link{join_to_spatial}}.}

\item{congress}{The number of the Congress. If \code{NULL}, it will be
guessed from the data.}

\item{projection}{If not provided, then the best state plane projection will
be guessed using the \code{\link[USAboundaries]{state_plane}} function from
the \code{USAboundaries} package. If \code{NULL}, then leaflet's default
Web Mercator projection will be used. To use a custom projection, provide a
projection/CRS object returned by the \code{\link[leaflet]{leafletCRS}}
function in the \code{leaflet} package.}

\item{congressional_boundaries}{Draw Congressional district boundaries in
addition to county boundaries?}

\item{cities}{Number of largest cities to draw. Pass \code{FALSE} to not draw
any cities.}

\item{width}{The width of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}

\item{height}{The height of the map in pixels or percentage. Passed on to
\code{\link[leaflet]{leaflet}}.}
}
\description{
Map elections data
}
\examples{
map_data <- get_county_map_data("meae.congressional.congress02.ct.county")
map_counties(map_data)

}
