% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{meae_candidates}
\alias{meae_candidates}
\alias{meae_congressional_counties}
\alias{meae_elections}
\alias{meae_maps_to_elections}
\alias{meae_maps}
\alias{meae_congress_counties_parties}
\alias{meae_congress_candidate_totals}
\alias{meae_congbio_elected}
\alias{nnv_name_authorities}
\title{Elections data from Mapping Early American Elections}
\format{Data frames}
\usage{
meae_candidates

meae_congressional_counties

meae_elections

meae_maps_to_elections

meae_maps

meae_congress_counties_parties

meae_congress_candidate_totals

meae_congbio_elected

nnv_name_authorities
}
\description{
Elections data from Mapping Early American Elections
}
\keyword{datasets}
